+++ changelog.md
@@ -2,6 +2,8 @@
 > **Note:**
 > Want to improve something? Check out the [format](#format)
 ## 11/28/2022
+- Update to latest discord.js
+## 11/28/2022
 - Add find the word!
 #### 11/27/2022
 - Add would you rather